Restructure Printer class
