+++ src/main.cpp
@@ -4,8 +4,7 @@
 #include "controller/input/Signal_Handler.h"
 
 int main() {
-  view::Printer printer;
-  printer.setVerboseLevel(Verbose::INFO);
+  view::Printer::setVerboseLevel(Verbose::INFO);
 
   controller::GameEngine engine;
   signal(SIGINT, Input::signal_callback_handler);
