--- src/view/PrintConstants.h
@@ -1,16 +1,0 @@
-//
-// Created by nolasco on 03/06/20.
-//
-
-#pragma once
-
-#include <string>
-
-namespace view {
-
-  constexpr auto separator1 = "################################################################";
-  constexpr auto separator2 = "----------------------------------------------------------------";
-  constexpr auto separator3 = "________________________________________________________________";
-  constexpr auto separatorOpen = "<<<<<<<<<<<<<<<<<<<<<<<<<<<<";
-  constexpr auto separatorClose = ">>>>>>>>>>>>>>>>>>>>>>>>>>>>";
-}
