+++ src/view/entity/CMakeLists.txt
@@ -1,0 +1,8 @@
+target_sources(${PROJECT_NAME}
+    PRIVATE
+        # Sources
+        ${CMAKE_CURRENT_SOURCE_DIR}/StreamConverter.cpp
+    PUBLIC
+        # Headers
+        ${CMAKE_CURRENT_SOURCE_DIR}/StreamConverter.h
+)
