+++ src/view/entity/StreamConverter.cpp
