+++ src/view/entity/StreamConverter.h
