+++ src/view/include/CMakeLists.txt
@@ -1,8 +1,8 @@
 target_sources(${PROJECT_NAME}
     PRIVATE
         # Sources
-        ${CMAKE_CURRENT_SOURCE_DIR}/StreamConverter.cpp
     PUBLIC
         # Headers
-        ${CMAKE_CURRENT_SOURCE_DIR}/StreamConverter.h
+        ${CMAKE_CURRENT_SOURCE_DIR}/Verbose.h
+        ${CMAKE_CURRENT_SOURCE_DIR}/PrintConstants.h
 )
