+++ src/view/include/PrintConstants.h
@@ -1,0 +1,16 @@
+//
+// Created by nolasco on 03/06/20.
+//
+
+#pragma once
+
+#include <string>
+
+namespace view {
+
+constexpr auto separator1 = "################################################################";
+constexpr auto separator2 = "----------------------------------------------------------------";
+constexpr auto separator3 = "________________________________________________________________";
+constexpr auto separatorOpen = "<<<<<<<<<<<<<<<<<<<<<<<<<<<<";
+constexpr auto separatorClose = ">>>>>>>>>>>>>>>>>>>>>>>>>>>>";
+} // namespace view
